fixup: replace :not(:first-child) selector
The :not pseudo-element is not supported in certain browsers, so we fall
back to setting all immediate children to "display: none" and use a more
specific selector with :first-child to set the first child to "display:
inline". Naturally, this fails to work on elements that are naturally
"display: block" or "display: inline-block" or "flex" or whatever, but
we can arguably create more specific classes for each of those cases.