refactor json-schema validators into OOP interface
The object-oriented interface is a lot more extensible, and uses a
builder pattern for setting options in the validators.

So far, the only option that has been added is to indicate to the
validator that a specific json schema should be used. Before, the json
schema to use was inferred based on the request path. However, for
API endpoints with variable paths (e.g. paths including user ids), this
inference will fail with a file not found error. Schema location
inference is still possible, and it's the default.