upgrade database
Add an `Administrator` table, representing another account type.
Accounts of this type are meant for bots and analytics tools that need
unrestricted access to the API.

Refactor common data in `Employee` and `Manager` into a `Human` table.
The `Human`/`Employee` and `Human`/`Manager` links are both one-to-one,
i.e. there is a unique constraint on the `human_id` column in both the
`Manager` and `Employee` table; lookups for identifying the account
owning a given `Human` should be fast.

A unique constraint is added on the `username` column in the `Login`
table.
table.