initialize a redis instance in the application core
The only use of Redis so far is in the redis-backed server-side sessions
code, so it's fine, in principle, to have that module create its own
redis instance. However, when we'll want to use redis in other aspects
of the application, we'll need to have a globally accessibly redis
instance.