implement real to_dict method for Administrator
Administrators have no interesting properties, so I held off on
implementing a to_dict method for them. Furthermore, API endpoints
should never be providing Administrator instances. However, I could see
it being useful in the future.

The implementation as is just provides the `id` column as well as the
`username` of the joined Login instance.