extract job_status_name from job_status spec
`job_status_name` is merely the enum of names for all job statuses.
These names are going to be used in API endpoints when the client wants
to specify a job status, so it's worthwhile to extract it from the
`job_status` specification, which is an object with a handful of other
properties.