change print statements to logger calls in api_spec
However, nothing gets printed and nobody is sure why. It could be due to
the fact that since `api_spec` does most of its work on import that the
logger calls are silently failing.

A potential strategy to remedy this might be a partial rewrite of
`api_spec` as a flask extension and fully decoupling it from the `auth`
module.