reorder api validators to avoid 500 errors when 400 errors occur
By running the request validator over the response validator, we ensure
that any failure responses generated by the request validator won't be
validated by the response validator. The upshot of this is that if the
request validator fails, and generates a 400 response, then the API
endpoint doesn't need to declare that a 400 errors can be generated.
Normally, if an API endpoint generates an undeclared response code, the
response validator will throw a ServerValidationError and ultimately
generate a 500 error.