create api endpoints for querying serverside enums
Turns out that we maintain quite a few enumerations of sorts in the
database. The clientside needs this for various forms, so thess api
endpoints were needed.

The following can now be queried by any authenticated client:

* languages
* cities
* states/provinces
* countries
* genders

The clientside BusinessService is updated to include functions for
performing those queries on the clientside.