fix up type-related bug in query string handling
The query string arguments are all taken as strings by Flask, which
means that a JSON validator expecting an integer will fail on that
input. To really solve this would require writing our own query string
parser that performs a JSON-like type inference, at least for integers.
Instead, we opt for declaring query_string parameters all as strings and
performing the parsing within the view function.