create ListingFormService and refactor new-listing view to use it
ListingFormService is a service that handles all the logic of forms that
describe listings, and it supports submitting forms to either create new
listings or update existing ones.

Specifically, a ListingForm object represents a form that describes a
listing, and ListingFormService contains a function for submitting these
objects to the backend.

To convert from the flat format reflected by the formly model into the
format expected by the backend, we implement the Adapter pattern with
the ListingFormAdapter.

The NewListingDetailsCtrl controller no longer needs to maintain all the
form state, since that's now handled by the ListingForm. Instead, the
controller instantiates a ListingForm object, which in turn introduces
an extra level of indirection to access the formly model. This required
modifying the new-listing template to account for this.

Finally, the BusinessApp module declaration was adjusted to include the
registration of the ListingFormService, and to inject the service into
the NewListingDetailsCtrl controller.