create edit-listing view
The view for editing listings is the same as the view for creating
listings up to some minor textual differences in headers and responses.

From the contracts menu, a listing can be edited by clicking on the Edit
button that is next to the delete button. These still need to be styled
properly, and ideally, the text should be replaced by glyphicons.

ListingEditCtrl controller uses ListingFormService in the same way that
NewListingDetailsCtrl does, i.e. by creating an instance of it and using
an extra level of indirection to deal with formly. The difference
between ListingEditCtrl and NewListingDetailsCtrl is that
ListingEditCtrl uses the third argument to ListingFormService's `submit`
function, namely a listing id. Passing this id makes ListingFormService
send a PATCH request, updating an existing listing rather than create a
new one.

There is a not-so-subtle bug having to do with position deletion.
Specifically, if the listing to edit is using a position that was
deleted, then the position select box remains empty and form submission
fails.