refactor Location model into FixedLocation and create Geolocation model
The Geolocation model represents a point on the surface of the Earth,
represented as a longitude and latitude pair. A FixedLocation represents
a legal address tied to a city, which also has a geographical location.
Hence, FixedLocation -> Geolocation. However, Employees track their
current locations, so Employee -> Geolocation as well.

This update also meant a change in terminology, to improve consistency.
The `home_location` and `location` columns of Employee and Business,
respectively, have both been renamed to `fixed_location`, as they refer
to rows from the FixedLocation table. As for the Employee's current
position, this is represented by a new `current_location` column, which
refers to a Geolocation row.