refactor collection API endpoints and use anti-XSRF cruft
The collection API endpoints used to return their arrays wrapped within
an object, which would complicate interactions with the Android
app. Instead, we make the collection endpoints return raw arrays.

This is however vulnerable to a flaw in JSON parsing that can allow a
malicious third-party to perform a cross-site request forgery attack by
intercepting the JSON data in the browser.

To remedy this, we prepend an unparseable cruft to every response JSON.
AngularJS automatically strips this cruft, so no work is required there.
However, Flask's jsonify function refuses to encode raw arrays due to
this vulnerability, so we write our own.

This also required a slight change in the response JSON schema
validator. In particular, it needs to strip the anti-XSRF marker prior
to parsing the JSON data.

On the client side, since we're returning raw arrays, the API interface
becomes a little bit simplier, since we no longer need to peel off a
layer of JavaScript object before getting to the array.