improve anti-XSRF to depend on HTTP headers
The Android application is unable to strip the anti-XSRF cruft that is
prepended to the JSON responses. To solve this, we check the User-agent
header of the request and make the app use a particular User-agent, in
particular "AhvodaApp". When this User-agent makes a request, the
anti-XSRF cruft is not prepended, which allows the app to properly parse
teh response body.