fix endpoints for employee arrival, departure, and application
First, the arrival and departure endpoints were attempting to store the
arrival/departure datetime on an attribute that didn't exist, which
(because we're using Python) isn't an error, but would instead just
result in the database not taking note of our changes.

Second, the endpoint to deal with applying needed some extra validation
logic to avoid cases when someone applies twice or when an employee is
already approved.