improve robustness of job-matcher script against database errors / races
Although this can't really happen in realistic scenarios, if a job is
pushed to the matcher's input queue and is then deleted, then when the
matcher tries to insert the relevant row to the JobMatch table, an
integrity check for a ForeignKey constraint will fail.

The script is now made to catch the IntegrityError and simply skip that
entry. The script also more gracefully handles KeyboardInterrupt by
omitting the stacktrace.