fix JobMatch SQLAlchemy model error involving the id column
JobMatch rows are identified by an id, but are also uniquely
identifiable by the tuple (Job.id, Employee.id). Both could be primary
keys, with the other merely having a unique constraint, so we
arbitrarily chose to have the tuple be the primary key.

Now this slightly messes up SQLAlchemy. The JobMatch model only declared
a unique and not null constraint on the id column, so when performing
inserts, an explicit `null` was used for the id column because
SQLAlchemy wants to be as explicit as possible unless it's told about
a server default. This would cause an integrity error upon inserting to
the JobMatch table.

This patch provides a server_default indication in the model, to avoid
SQLAlchemy using explicit nulls.