improve API spec properties and add applicants list to listing instance
Listing instances did not include a list of their applicants, which
significantly complicated the code on the client-side, necessitating
extra API calls to get that list. Instead, we can just bundle the
applicants list server-side.

Right now, no discretion is made about the applicants: anyone who can
access a listing instance will get the complete list of applicants. We
will need to look into filtering the applicants to those that the
authenticated user can see.

The jsonschema2pojo utility used to generate Java classes used some kind
of weird opaque heuristics for generating the class names. I thought it
was based on the file name, but no. Also, it did not include a package
statement in the generated files, which required some manual effort in
Android Studio afterwards. Now, by including a "javaType" property to
each definition object, jsonschema2pojo is capable of using exactly the
class name that I want, and it puts the data into a package.